Springcloud consul

+++ shorturl-springcloud/redirect-service/src/main/java/com/example/redirectservice/service/RedirectServiceImpl.java
@@ -4,10 +4,11 @@ import com.example.sharedentity.dao.ShortUrlDao;
 import com.example.sharedentity.entity.ShortUrl;
 import com.example.sharedentity.util.Base62Encoder;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.messaging.SubscribableChannel;
 import org.springframework.messaging.support.MessageBuilder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.*;
 
 @Service
 @Transactional(rollbackFor = Exception.class)
@@ -16,16 +17,14 @@ public class RedirectServiceImpl implements RedirectService {
     private ShortUrlDao shortUrlDao;
     @Autowired
     private ISendeService sendService;
+
     @Override
     public ShortUrl findShortUrlByToken(String token) throws Exception {
-        int shortUrlId = Base62Encoder.decode(token);
+        long shortUrlId = Base62Encoder.decode(token);
         ShortUrl shortUrl = shortUrlDao.findById(shortUrlId);
         sendService.send().send(MessageBuilder
                 .withPayload(shortUrlId)
                 .build());
-   //     shortUrl.setClicks(shortUrl.getClicks()+1);
-     //   shortUrlDao.saveAndFlush(shortUrl);
         return shortUrl;
     }
-
 }
